requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-dotenv